+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.2.0
+-----
+
+-   Send versionCode & applicationId
+-   Fix when -dontobfuscate is configured in proguard
+-   Fix when multiple variants have proguard enabled
+
 0.1.0
 -----